Support other UCI 'go' command args
Now supports:

* go movetime 1000
* go depth 5
* go infinite

Also passes movestogo to TimeStrategy but doesn't use it yet

+++ engine/src/uci/commands.rs
@@ -15,7 +15,7 @@ pub struct GoCmdArguments {
     pub winc: Option<u32>,
     pub binc: Option<u32>,
     pub movestogo: Option<u32>,
-    pub depth: Option<u32>,
+    pub depth: Option<u8>,
     pub nodes: Option<u32>,
     pub mate: Option<u32>,
     pub movetime: Option<u32>,
